Adds alpha version of /api/packages/:package_name
Responds with 200 to /api/packages/:package_name if package found and with 404 otherwise.
Verifies if the package has been found by parsing the position of the first line of the definition of a package
into a dictionary that includes the name of the package and the line in the file where it can be found.